已经都转到了BigDecimal，以后不用double了，放心计算小数了

+++ src/main/java/com/eg/testbitcoinpay/huobi/QueryPriceUtil.java
@@ -4,11 +4,13 @@ import com.alibaba.fastjson.JSON;
 import com.eg.testbitcoinpay.huobi.markettrade.MarketTradeResponse;
 import com.eg.testbitcoinpay.util.HttpUtil;
 
+import java.math.BigDecimal;
+
 /**
  * 查询币价
  */
 public class QueryPriceUtil {
-    public static double getBitcoinPrice() {
+    public static BigDecimal getBitcoinPrice() {
         String url = "https://api.huobi.com/market/trade?symbol=btcusdt";
         MarketTradeResponse marketTradeResponse = JSON.parseObject(HttpUtil.getProxy(url), MarketTradeResponse.class);
         return marketTradeResponse.getTick().getData().get(0).getPrice();
